LayoutTests/fast/scrolling/scroll-anchoring/heuristic-enabled-large-cumulative-adjustment-expected.txt


Testing anchoring with 10 of size 10 - expect anchoring true
PASS anchoringActive is true
PASS successfullyParsed is true

TEST COMPLETE
